@@ -1,17 +1,19 @@
 package general;
 
-import domain.Sucursal;
-import domain.impuestos.EI;
-import domain.impuestos.EO;
-import domain.impuestos.IVA;
-import domain.observers.ContadorDeVentas;
-import domain.observers.Deposito;
-import domain.observers.OficinaContable;
-import domain.productos.Producto;
-import domain.productos.TipoProducto;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+import ar.utn.ba.ddsi.garbarisi.models.entities.Sucursal;
+import ar.utn.ba.ddsi.garbarisi.models.entities.impuestos.EI;
+import ar.utn.ba.ddsi.garbarisi.models.entities.impuestos.EO;
+import ar.utn.ba.ddsi.garbarisi.models.entities.impuestos.IVA;
+import ar.utn.ba.ddsi.garbarisi.models.entities.observers.ContadorDeVentas;
+import ar.utn.ba.ddsi.garbarisi.models.entities.observers.Deposito;
+import ar.utn.ba.ddsi.garbarisi.models.entities.observers.OficinaContable;
+import ar.utn.ba.ddsi.garbarisi.models.entities.observers.adapters.AdapterSistemaContable;
+import ar.utn.ba.ddsi.garbarisi.models.entities.productos.Producto;
+import ar.utn.ba.ddsi.garbarisi.models.entities.productos.TipoProducto;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 public class ProductoTest {
 	private Sucursal sucursal;
@@ -22,8 +24,9 @@ public class ProductoTest {
 	private TipoProducto hogar;
 	private Producto celu;
 	private Producto estufa;
+	private AdapterSistemaContable adapterSistemaContable;
 
-	@Before
+	@BeforeEach
 	public void init(){
 		EI.setGananciasImpositivas(3.50);
 		EI.setFactorPrecioBase(4.0);
@@ -41,18 +44,15 @@ public class ProductoTest {
 		hogar = new TipoProducto();
 		hogar.agregarImpuestos(new EI(), new IVA());
 
+		this.adapterSistemaContable = Mockito.mock(AdapterSistemaContable.class);
 		this.contador = new ContadorDeVentas();
-		this.oficinaContable = new OficinaContable();
+		this.oficinaContable = new OficinaContable(adapterSistemaContable);
 		this.deposito = new Deposito();
-		this.celu = new Producto("Samsumng S20 PRO", 120000.00, electronico);
-		this.estufa = new Producto("Estufa tiro balanceado Orbis", 15000.00, hogar);
+		this.celu = new Producto(electronico,  120000.00, "Samsumng S20 PRO");
+		this.estufa = new Producto(hogar, 15000.00, "Estufa tiro balanceado Orbis");
 
 		this.sucursal = new Sucursal("Casa Central");
 		this.sucursal.agregarProductos(this.celu, this.estufa);
-
-		this.deposito = new Deposito();
-		this.oficinaContable = new OficinaContable();
-		this.contador = new ContadorDeVentas();
 
 		this.sucursal.agregarObservers(deposito, oficinaContable, contador);
 	}
@@ -60,18 +60,19 @@ public class ProductoTest {
 	@Test
 	public void oficinaContableSeEnteraDeLaVentaDelCelu(){
 		this.sucursal.vende(this.celu);
-		Assert.assertTrue(this.oficinaContable.getLibroDiario() > this.celu.getPrecioBase());
+		Assertions.assertTrue(this.oficinaContable.getTotalLibroDiario() > this.celu.getPrecioBase());
+		Mockito.verify(adapterSistemaContable).notificar(this.celu);
 	}
 
 	@Test
 	public void contadorDeVentasSeEnteraDeLaVentaDelCelu(){
 		this.sucursal.vende(this.celu);
-		Assert.assertEquals(1, this.contador.getCantidadVentas());
+		Assertions.assertEquals(1, this.contador.getCantidadVentas());
 	}
 
 	@Test
 	public void depositoPreparaCelu(){
 		this.sucursal.vende(this.celu);
-		Assert.assertTrue(this.deposito.getProductosAPreparar().contains(this.celu));
+		Assertions.assertTrue(this.deposito.getProductosAPreparar().contains(this.celu));
 	}
 }
