@@ -1,26 +1,22 @@
-package domain.productos;
+package ar.utn.ba.ddsi.garbarisi.models.entities.productos;
 
-import domain.impuestos.Impuesto;
+import ar.utn.ba.ddsi.garbarisi.models.entities.impuestos.Impuesto;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+@Getter
 public class TipoProducto {
+    @Setter
     private String descripcion;
     private List<Impuesto> impuestos;
 
     public TipoProducto(){
         super();
         this.impuestos = new ArrayList<>();
-    }
-
-    public List<Impuesto> getImpuestos() {
-        return impuestos;
-    }
-
-    public String getDescripcion() {
-        return descripcion;
     }
 
     public void agregarImpuestos(Impuesto ... impuestos){
